pyproject.toml
[project]
name = "yolk-demo"
version = "0.1.0"
description = "AI Sales Coach - Backend Training Project (Yolk Interview Prep)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic[email]>=2.10.0",
    "pydantic-settings>=2.6.0",
    "redis>=5.2.0",
    "httpx>=0.28.0",
    "openai>=1.57.0",
    "anthropic>=0.39.0",
    "faststream[rabbit]>=0.5.0",
    "opentelemetry-api>=1.28.0",
    "opentelemetry-sdk>=1.28.0",
    "opentelemetry-instrumentation-fastapi>=0.49b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.49b0",
    "opentelemetry-instrumentation-httpx>=0.49b0",
    "opentelemetry-exporter-otlp>=1.28.0",
    "websockets>=14.0",
    "structlog>=24.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "basedpyright>=1.22.0",
    "faker>=33.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/yolk"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E", "W", "F", "I", "N", "UP", "B", "A",
    "ASYNC", "S", "T20", "RET", "SIM", "TCH", "RUF",
]
ignore = ["S101", "B008"]

[tool.ruff.lint.per-file-ignores]
"scripts/*" = ["T201", "ASYNC250"]

[tool.ruff.lint.isort]
known-first-party = ["yolk"]

[tool.basedpyright]
pythonVersion = "3.12"
typeCheckingMode = "strict"
reportMissingTypeStubs = false
reportUnknownMemberType = false
reportUnknownArgumentType = false
reportUnknownVariableType = false
reportUnknownParameterType = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]
